[exporter/kafka] move kafka configures authentication to internal pkg
