+++ .github/ISSUE_TEMPLATE/bug_report.yaml
@@ -110,6 +110,7 @@ body:
       - internal/filter
       - internal/k8sconfig
       - internal/k8stest
+      - internal/kafka
       - internal/kubelet
       - internal/metadataproviders
       - internal/sharedcomponent
