+++ internal/kafka/Makefile
@@ -1,0 +1,1 @@
+include ../../Makefile.Common
