+++ internal/kafka/metadata.yaml
@@ -1,0 +1,3 @@
+status:
+  codeowners:
+    active: [pavolloffay, MovieStoreGuy]
