+++ receiver/kafkareceiver/config_test.go
@@ -15,6 +15,7 @@ import (
 	"go.opentelemetry.io/collector/confmap/confmaptest"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/kafkaexporter"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/kafka"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/kafkareceiver/internal/metadata"
 )
 
@@ -38,7 +39,7 @@ func TestLoadConfig(t *testing.T) {
 				ClientID:      "otel-collector",
 				GroupID:       "otel-collector",
 				InitialOffset: "latest",
-				Authentication: kafkaexporter.Authentication{
+				Authentication: kafka.Authentication{
 					TLS: &configtls.TLSClientSetting{
 						TLSSetting: configtls.TLSSetting{
 							CAFile:   "ca.pem",
@@ -70,7 +71,7 @@ func TestLoadConfig(t *testing.T) {
 				ClientID:      "otel-collector",
 				GroupID:       "otel-collector",
 				InitialOffset: "earliest",
-				Authentication: kafkaexporter.Authentication{
+				Authentication: kafka.Authentication{
 					TLS: &configtls.TLSClientSetting{
 						TLSSetting: configtls.TLSSetting{
 							CAFile:   "ca.pem",
